# openapi.yaml
openapi: 3.1.0
info:
  title: simpletickr API
  version: 0.1.0
  description: Simple portfolio tracker for ETFs, stocks, and crypto.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - cookieAuth: []
...
paths:
...
  /portfolios:
    get:
      tags: [Portfolios]
      operationId: listPortfolios
      summary: List all portfolios
      responses:
        "200":
          description: List of portfolios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Portfolio"
    post:
      tags: [Portfolios]
      operationId: createPortfolio
      summary: Create a portfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioRequest"
      responses:
        "201":
          description: Portfolio created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
  /portfolios/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Portfolios]
      operationId: getPortfolio
      summary: Get a portfolio by ID
      responses:
        "200":
          description: Portfolio found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
        "404":
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [Portfolios]
      operationId: updatePortfolio
      summary: Update a portfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioRequest"
      responses:
        "200":
          description: Portfolio updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
        "404":
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Portfolios]
      operationId: deletePortfolio
      summary: Delete a portfolio
      responses:
        "204":
          description: Portfolio deleted
        "404":
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /portfolios/{id}/realized-gains:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Portfolios]
      operationId: getRealizedGains
      summary: Compute realized gains for a portfolio
      parameters:
        - name: method
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/RealizationMethod"
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Realized gains report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RealizedGainsReport"
        "404":
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /portfolios/{id}/holdings:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Portfolios]
      operationId: getHoldings
      summary: Get computed holdings for a portfolio
      responses:
        "200":
          description: Holdings for the portfolio
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Holding"
        "404":
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /portfolios/{id}/account-allocation:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Portfolios]
      operationId: getAccountAllocation
      summary: Get current market value per account holding assets in this portfolio
      responses:
        "200":
          description: Account allocation for the portfolio
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountAllocation"
        "404":
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /portfolios/{id}/valuation-summary:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Portfolios]
      operationId: getPortfolioValuationSummary
      summary: Get portfolio-level market value and unrealized gain, excluding holdings with unavailable prices
      responses:
        "200":
          description: Valuation summary for the portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioValuationSummary"
        "404":
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /portfolios/{id}/value-history:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Portfolios]
      operationId: getPortfolioValueHistory
      summary: Get daily total portfolio value over a date range
      parameters:
        - name: from
          in: query
          required: false
          description: "Start date (inclusive). Defaults to the portfolio's oldest transaction date."
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Time series of daily portfolio value and invested amount in base currency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioValueHistory"
        "404":
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /portfolios/{id}/sync-prices:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    post:
      tags: [Portfolios]
      operationId: syncPortfolioPrices
      summary: Backfill price history for all listings in a portfolio from the oldest transaction date
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
        "404":
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /settings:
    get:
      tags: [Settings]
      operationId: getSettings
      summary: Get user settings
      responses:
        "200":
          description: User settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      tags: [Settings]
      operationId: updateSettings
      summary: Update user settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Settings"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: SESSION
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    AccountType:
      type: string
      enum:
        - BROKERAGE
        - CRYPTO
        - BANK
        - RETIREMENT
        - OTHER
...
    Portfolio:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    PortfolioRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
...
    SyncResult:
      type: object
      required: [synced, failed]
      properties:
        synced:
          type: integer
        failed:
          type: integer
    RealizationMethod:
      type: string
      enum:
        - FIFO
        - AVERAGE_COST
    RealizedGainEntry:
      type: object
      required:
        - assetId
        - ticker
        - currency
        - date
        - quantity
        - proceeds
        - buyFees
        - sellFees
        - costBasis
        - gain
        - lots
      properties:
        assetId:
          type: integer
          format: int64
        ticker:
          type: string
        currency:
          type: string
          description: Local currency of the sell transaction.
          example: EUR
        date:
          type: string
          format: date
        quantity:
          type: number
          format: double
        proceeds:
          type: number
          format: double
          description: In local currency.
        buyFees:
          type: number
          format: double
          description: In local currency.
        sellFees:
          type: number
          format: double
          description: In local currency.
        costBasis:
          type: number
          format: double
          description: In local currency.
        gain:
          type: number
          format: double
          description: In local currency. Do not sum across currencies.
        tradeId:
          type: integer
          format: int64
          description: Present when this sell was one leg of a crypto-to-crypto swap.
        receivedTicker:
          type: string
          description: Ticker of the asset received in the swap. Present when tradeId is set.
        lots:
          type: array
          description: "Which acquisition lot(s) generated this gain. FIFO only — always empty for AVERAGE_COST, which has no discrete lots."
          items:
            $ref: "#/components/schemas/RealizedGainLot"
    RealizedGainLot:
      type: object
      description: "One acquisition lot's contribution to a RealizedGainEntry. costBasis/buyFees/quantity sum across all lots on an entry to that entry's own fields."
      required:
        - acquisitionDate
        - quantity
        - pricePerUnit
        - buyFees
        - costBasis
      properties:
        acquisitionDate:
          type: string
          format: date
        quantity:
          type: number
          format: double
        pricePerUnit:
          type: number
          format: double
          description: In local currency.
        buyFees:
          type: number
          format: double
          description: In local currency.
        costBasis:
          type: number
          format: double
          description: "In local currency. quantity * pricePerUnit + buyFees, for this lot's share only."
    RealizedGainsReport:
      type: object
      required:
        - method
        - from
        - to
        - entries
        - byCurrency
      properties:
        method:
          $ref: "#/components/schemas/RealizationMethod"
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        entries:
          type: array
          items:
            $ref: "#/components/schemas/RealizedGainEntry"
        byCurrency:
          type: object
          description: Per-currency totals. Safe to display independently; never sum across currencies.
          additionalProperties:
            $ref: "#/components/schemas/CurrencyTotal"
    CurrencyTotal:
      type: object
      required:
        - currency
        - tradeCount
        - totalProceeds
        - totalCostBasis
        - totalGain
      properties:
        currency:
          type: string
        tradeCount:
          type: integer
        totalProceeds:
          type: number
          format: double
        totalCostBasis:
          type: number
          format: double
        totalGain:
          type: number
          format: double
    Settings:
      type: object
      required:
        - baseCurrency
      properties:
        baseCurrency:
          type: string
          example: EUR
    Holding:
      type: object
      description: Asset-level holdings view with per-listing drill-down. All monetary values are in baseCurrency.
      required:
        - assetId
        - assetName
        - baseCurrency
        - totalQuantity
        - listings
      properties:
        assetId:
          type: integer
          format: int64
        assetName:
          type: string
        baseCurrency:
          type: string
          example: EUR
        totalQuantity:
          type: number
          format: double
        avgCostBasisBase:
          type: number
          format: double
          nullable: true
          description: FX-normalized weighted average cost per unit. Null only when every listing lacks an FX rate.
        totalCostBase:
          type: number
          format: double
          nullable: true
          description: FX-normalized total cost, summing whichever listings have an FX rate. Null only when every listing lacks one.
        marketValueBase:
          type: number
          format: double
          nullable: true
          description: Summed across whichever listings have a price. Null only when no listing has one.
        unrealizedPnlBase:
          type: number
          format: double
          nullable: true
        unrealizedPnlPct:
          type: number
          format: double
          nullable: true
        listings:
          type: array
          items:
            $ref: "#/components/schemas/ListingHolding"
    PortfolioValuationSummary:
      type: object
      description: Portfolio-level market value and unrealized gain, in baseCurrency. Explicitly reports which holdings were excluded rather than leaving consumers to infer it from a null.
      required:
        - totalCostBase
        - excludedHoldingCount
        - excludedHoldingNames
      properties:
        totalCostBase:
          type: number
          format: double
          description: Always fully known — cost basis doesn't require current market price data.
        totalMarketValueBase:
          type: number
          format: double
          nullable: true
          description: Summed across whichever holdings have a price. Null only if no holding in the portfolio has one.
        totalUnrealizedPnlBase:
          type: number
          format: double
          nullable: true
        totalUnrealizedPnlPct:
          type: number
          format: double
          nullable: true
          description: Computed against the full totalCostBase, including excluded holdings — a deliberately conservative (not inflated) percentage.
        excludedHoldingCount:
          type: integer
        excludedHoldingNames:
          type: array
          items:
            type: string
          description: Asset names excluded from totalMarketValueBase/totalUnrealizedPnl* because no listing of that asset has a price.
    AccountAllocation:
      type: object
      description: Current market value held in one account, in baseCurrency. No cost basis — transfers move custody with no price, so cost can't be attributed to a specific account.
      required:
        - accountId
        - accountName
        - accountType
        - baseCurrency
      properties:
        accountId:
          type: integer
          format: int64
        accountName:
          type: string
        accountType:
          $ref: "#/components/schemas/AccountType"
        baseCurrency:
          type: string
          example: EUR
        marketValueBase:
          type: number
          format: double
          nullable: true
          description: Summed across whichever holdings in this account have a price. Null when none do.
    ListingHolding:
      type: object
      required:
        - listingId
        - ticker
        - currency
        - quantity
        - avgCostLocal
        - totalCostLocal
      properties:
        listingId:
          type: integer
          format: int64
        exchange:
          type: string
          nullable: true
        ticker:
          type: string
        currency:
          type: string
          example: USD
        quantity:
          type: number
          format: double
        avgCostLocal:
          type: number
          format: double
        totalCostLocal:
          type: number
          format: double
        marketValueLocal:
          type: number
          format: double
          nullable: true
        fxRate:
          type: number
          format: double
          nullable: true
          description: Rate used for conversion (1 baseCurrency = fxRate listingCurrency). Null when currencies match.
        marketValueBase:
          type: number
          format: double
          nullable: true
...
    PortfolioValuePoint:
      type: object
      required: [date]
      properties:
        date:
          type: string
          format: date
        value:
          type: number
          format: double
          nullable: true
          description: >
            Market value in base currency on this date. Null when any open holding lacks a
            price or FX rate (partial sums are not returned to avoid understating the portfolio).
        invested:
          type: number
          format: double
          nullable: true
          description: >
            Net cumulative cash invested in base currency (BUY costs minus SELL proceeds),
            using the FX rate recorded on each transaction.
    PortfolioValueHistory:
      type: object
      required: [baseCurrency, points]
      properties:
        baseCurrency:
          type: string
          example: EUR
        points:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioValuePoint"
...